+++ decidim-proposals/app/commands/decidim/proposals/admin/update_participatory_text.rb
@@ -3,9 +3,8 @@
 module Decidim
   module Proposals
     module Admin
-      # A command with all the business logic when an admin imports proposals from
-      # a participatory text.
-      class PublishParticipatoryText < UpdateParticipatoryText
+      # A command with all the business logic when an admin updates participatory text proposals.
+      class UpdateParticipatoryText < Rectify::Command
         # Public: Initializes the command.
         #
         # form - A PreviewParticipatoryTextForm form object with the params.
@@ -23,7 +22,6 @@ module Decidim
           transaction do
             @failures = {}
             update_contents_and_resort_proposals(form)
-            publish_drafts
           end
 
           if @failures.any?
@@ -37,9 +35,14 @@ module Decidim
 
         attr_reader :form
 
-        def publish_drafts
-          Decidim::Proposals::Proposal.where(component: form.current_component).drafts.find_each do |proposal|
-            add_failure(proposal) unless proposal.update(published_at: Time.current)
+        def update_contents_and_resort_proposals(form)
+          form.proposals.each do |prop_form|
+            proposal = Decidim::Proposals::Proposal.where(component: form.current_component).find(prop_form.id)
+            proposal.set_list_position(prop_form.position) if proposal.position != prop_form.position
+            proposal.title = prop_form.title
+            proposal.body = prop_form.body if proposal.participatory_text_level == Decidim::Proposals::ParticipatoryTextSection::LEVELS[:article]
+
+            add_failure(proposal) unless proposal.save
           end
           raise ActiveRecord::Rollback if @failures.any?
         end